requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\requirements\requirements.in'
#
beautifulsoup4==4.9.3
    # via -r .\requirements\requirements.in
lxml==4.6.2
    # via -r .\requirements\requirements.in
markdown-generator==0.1.3
    # via -r .\requirements\requirements.in
pyyaml==5.4.1
    # via -r .\requirements\requirements.in
soupsieve==2.1
    # via
    #   -r .\requirements\requirements.in
    #   beautifulsoup4
tqdm==4.56.0
    # via -r .\requirements\requirements.in
